Making tests use tcp connection to broker as much as possible.
Discovered a bug with closed Worker sockets not working after the worker had been closed and then re-opened. The root cause appears to be zmq blacklisting closed sockets by their identity. I realize that the worker can be meaningfully named so clobbering its the name might not be a good thing in general, but I don't know of another workaround at the moment.