fix: Failing CI
The CI was failing because the build was relying on specific version
branches to exist.
The development teams of both projects --- ZeroMQ and libsodium ---
created tags for each version and removed the branches.
The branch names were removed from the env vars and new env vars
containing the actual software versions were introduced.
After cloning the code, a separate checkout command is executed. This
command carries out successfully.

Signed-off-by: Adrian Oprea <adrian@oprea.rocks>